refactor(macos): extract items-query infra from AppModel into AppModel+ItemsQuery.swift
Continues the AppModel god-object split. Relocates the `/Items` query
construction + JSON→model parsing infra — buildItemsQuery (the authenticated
GET /Items request builder) and the static parsers that turn the
{ Items: [...] } envelope into typed Album/Artist/Track values
(parse*FromItems + the per-DTO decoders) — into a new `extension AppModel` file.

Pure relocation: the 322 moved lines are byte-identical to what was removed.
AppModel.swift drops 4,466 -> 4,143 lines. No stored state in the block.

Four access promotions (private static -> internal static): parseAlbumsFromItems,
parseAlbumsWithPlayCounts, parseLatestAlbumsWithDates, parseArtistsFromItems —
the parsers the fetch* helpers (staying in AppModel.swift) call cross-file. The
per-DTO decoders (artistFromDTO / albumFromDTO / trackFromDTO) and
parseTracksFromItems (already internal) stay as-is; buildItemsQuery was already
internal.

Verified: clean rebuild (rm -rf macos/.build && swift build) + swift test
(889 tests) green.
